@@ -1,4 +1,4 @@
-package com.project;
+package com.project.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
